refactor(layout): 네비게이션바, 친구선택박스, 선물선택박스, 광고 배너컴포넌트를 분리해줬습니다
